+++ sample/src/com/manuelpeinado/multichoiceadapter/demo/ActivityInfo.java
@@ -15,12 +15,14 @@
  */
 package com.manuelpeinado.multichoiceadapter.demo;
 
+import android.app.Activity;
+
 public class ActivityInfo {
-    public ActivityInfo(Class<? extends BaseActivity> activityClass, int titleResourceId) {
+    public Class<? extends Activity> activityClass;
+    public int titleResourceId;
+
+    public ActivityInfo(Class<? extends Activity> activityClass, int titleResourceId) {
         this.activityClass = activityClass;
         this.titleResourceId = titleResourceId;
     }
-
-    public Class<? extends BaseActivity> activityClass;
-    public int titleResourceId;
 }